requirements.txt
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0
django-grappelli==3.0.3
django-modeltranslation==0.18.7
django-phonenumber-field==6.4.0
django-zipfile==0.4.0
phonenumbers==8.12.54
Pillow==9.2.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.2
